Move Google Quickdraw renderer and aggregator to examples
Only reusable renderers and aggregators should be in the main source code
