+++ examples/google-quickdraw-cover-aggregator-worker.js
