+++ src/aggregator.js
@@ -1,4 +1,3 @@
 export { default as createMatrixPreviewAggregator } from './matrix-preview-aggregator';
 export { default as createMatrixCoverAggregator } from './matrix-cover-aggregator';
-export { default as createQuickDrawCoverAggregator } from './quickdraw-cover-aggregator';
 export { default as createRepresentativeAggregator } from './representative-aggregator';
