+++ src/renderer.js
@@ -2,5 +2,3 @@ export { default as createImageRenderer } from './image-renderer';
 export { default as createMatrixRenderer } from './matrix-renderer';
 export { default as createRepresentativeRenderer } from './representative-renderer';
 export { default as createSvgRenderer } from './svg-renderer';
-export { default as createQuickDrawRenderer } from './quickdraw-renderer';
-export { default as createQuickDrawCoverRenderer } from './quickdraw-cover-renderer';
